requirements.txt
streamlit
langchain-openai
langchain-huggingface
pdf2image
Pillow
pandas